requirements.txt
fastapi
uvicorn
psycopg2-binary
pydantic
PyJWT
bcrypt
SQLAlchemy
email-validator
pydantic-settings>=2.9.1
passlib[bcrypt]
